minor type debug

+++ src/db/models/debt.ts
@@ -1,4 +1,5 @@
 import Mongoose from 'mongoose'
+import { Decimal128 } from 'bson';
 
 export async function getDebts(id: number) {
     return [];
@@ -7,7 +8,7 @@ export async function getDebts(id: number) {
 type Debt = {
     from_id: number
     to_id: number
-    amount: BigInt
+    amount: Decimal128
     currency: string
 }
 
@@ -16,6 +17,6 @@ type DebtDoc = Mongoose.Document & Debt;
 const DebtModel = Mongoose.model<DebtDoc>('Debt', new Mongoose.Schema({
     from_id: { type: Number, required: true },
     to_id: { type: Number, required: true },
-    amount: { type: BigInt, required: true },
+    amount: { type: Decimal128, required: true },
     currency: { type: String, required: true }
 }));
